+++ relay-server/src/managed/managed/test.rs
@@ -105,7 +105,7 @@ impl ManagedTestHandle {
 
     /// Asserts a single emitted outcome.
     #[track_caller]
-    pub fn assert_outcome(&mut self, outcome: &Outcome, category: DataCategory, quantity: u32) {
+    pub fn assert_outcome(&mut self, outcome: &Outcome, category: DataCategory, quantity: u64) {
         match self.outcomes.try_recv() {
             Ok(next) => {
                 assert_eq!(&next.outcome, outcome);
@@ -127,7 +127,7 @@ impl ManagedTestHandle {
     /// A shorthand for [`Self::assert_outcome`] with a [`Outcome::Invalid`] /
     /// [`DiscardReason::Internal`].
     #[track_caller]
-    pub fn assert_internal_outcome(&mut self, category: DataCategory, quantity: u32) {
+    pub fn assert_internal_outcome(&mut self, category: DataCategory, quantity: u64) {
         self.assert_outcome(
             &Outcome::Invalid(DiscardReason::Internal),
             category,
